+++ Engine/RedTriangle.h
@@ -5,19 +5,19 @@
 class RedTriangle
 {
 public:
-	void Init(int in_x, int in_y, int in_vx, int in_vy);
+	void Init(float in_x, float in_y, float in_vx, float in_vy);
 	void Draw(Graphics& gfx) const;
 	void Movement();
 	void collection(const Triangle& triangle);
 	bool collect(const Triangle& triangle);
 	bool Collected() const;
 private:
-	int x;
-	int y;
-	int vx;
-	int vy;
-	static constexpr int width = 10;
-	static constexpr int height = 10;
+	float x;
+	float y;
+	float vx;
+	float vy;
+	static constexpr float width = 10.0f;
+	static constexpr float height = 10.0f;
 	bool collected = false;
 };
